Acquire the Hamburg master-city stack, and emit metric UV1 for facade materials
Two things, both aimed at the same failure: a correct inner-city block standing
on a green field with no streets.

Acquisition. Buildings were the only thing acquired, so buildings were the only
thing that could appear. The manifest now covers the rest and it is resolved
live from the Transparenzportal CKAN API rather than read off a page:

  terrain     DGM 1, 1 m grid, 2.95 GB    the ground the LoD3 was placed on but
                                          which the building package omits
  roads       HH-SIB network, GML         the seed's road list was empty, so the
                                          generator stamped nothing and the
                                          theme painted the Altstadt as grass
  cadastre    ALKIS selected data, 0.54 GB  pavement vs carriageway vs courtyard
  orthophoto  DOP20 via cached WMS        deliberately not bulk: an epoch is
                                          15 GB citywide for a few km2 of need

All dl-de/by-2-0. tools/fetch_hamburg.py walks the plan in the order that makes
a block look right soonest and stops on a disk budget rather than after it.

Found on the way: the fetcher skipped every building package and reported
success. `Area.url` is a property, so spreading `__dict__` into the plan omitted
it, and the guard was `if not url: continue` with nothing printed. Both halves
fixed -- the plan names the url, and a plan entry without one is now a reported
failure. A silent skip that reports success is the worst shape a bug can have
here.

UV1. Following the facade-mapping guidance, the textured backend now emits a
second texture channel in wall-local *metres*, alongside UV0 passed through
untouched. Normalising by wall width and height was the obvious choice and is
wrong: it makes one brick course span a shed and a warehouse identically. Metric
coordinates let the shader divide by a real material repeat.

The frame is oriented, not projected. Masonry is directional -- courses run
along the wall and stack toward the sky -- so v is world up and u is horizontal
along the face. A triplanar projection rotates brick at a corner, which reads as
fake instantly.

Also written: a surface index beside the model, carrying surface and building
ids, the wall frame and metric extent per face. That is what makes a mask
detected in UV0 mappable back onto a named wall in 3D; none of it survives
inside a glTF, and dropping the ids at the conversion boundary would make the
appearance work impossible afterwards.

Rathaus block re-exported: 51,140 triangles, 44 textures, 15,224 surfaces
indexed.

250 tests pass, 4 new covering metric scale, up-axis, corner orientation and
the index.

+++ src/lidarworld/reconstruct/mesh.py
@@ -37,7 +37,19 @@ class MeshBuilder:
         if q == 0:
             return
         base = self._vertex_count
-        self.positions.append(corners.reshape(-1, 3).astype(np.float32))
+        # float64, deliberately, and the only array here that is.
+        #
+        # These are projected coordinates: a UTM northing is about 5.93e6, where
+        # float32 resolves 0.5 m, and an easting about 5.7e5, where it resolves
+        # 0.0625 m. Casting on the way in snapped every vertex to that grid --
+        # measured on a Hamburg block, the smallest distinct northing in the
+        # exported buffer was exactly 0.5 m against 0.0625 m in easting. The
+        # result is an anisotropic sawtooth on any wall not aligned to the axes,
+        # which is every wall in a rotated street grid, and it moved no metric.
+        #
+        # Precision has to be kept until something recentres. Backends do that
+        # and cast there, where the local range is metres and float32 is ample.
+        self.positions.append(corners.reshape(-1, 3).astype(np.float64))
         if normal.ndim == 1:
             self.normals.append(np.repeat(normal[None, :], q * 4, axis=0).astype(np.float32))
         else:
@@ -58,7 +70,7 @@ class MeshBuilder:
     def finalize(self) -> dict[str, np.ndarray]:
         if not self.positions:
             return {
-                "positions": np.zeros((0, 3), np.float32), "normals": np.zeros((0, 3), np.float32),
+                "positions": np.zeros((0, 3), np.float64), "normals": np.zeros((0, 3), np.float32),
                 "uv": np.zeros((0, 2), np.float32), "ctx": np.zeros(0, np.uint32),
                 "role": np.zeros(0, np.uint8), "node": np.zeros(0, np.uint32),
                 "indices": np.zeros((0, 3), np.uint32),
